added privacy policy + some routing changes
