+++ src/Routes.js
@@ -1,0 +1,15 @@
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import PrivacyPolicy from "./pages/PrivacyPolicy";
+import HomePage from "./pages/HomePage";
+
+export const SiteRoutes = () => {
+    return (
+        <Router>
+            <Routes>
+                <Route exact path="/PrivacyPolicy" element={<PrivacyPolicy />} />
+                <Route exact path="/" element={<HomePage />} />
+            </Routes>
+        </Router>
+    );
+}
+export default SiteRoutes;
